+++ acp/main_module.php
@@ -59,6 +59,7 @@ class main_module
 			$config->set('primelinks_inlink_regex', $request->variable('primelinks_inlink_regex', ''));
 			$config->set('primelinks_exlink_regex', $request->variable('primelinks_exlink_regex', ''));
 			$config->set('primelinks_skip_prefix_regex', $request->variable('primelinks_skip_prefix_regex', ''));
+			$config->set('primelinks_inlink_use_titles', $request->variable('primelinks_inlink_use_titles', false));
 
 			$log = $phpbb_container->get('log');
 			$log->add('admin', $user->data['user_id'], $user->ip, 'ACP_PRIMELINKS_SETTINGS_LOG');
@@ -89,6 +90,7 @@ class main_module
 			'PRIMELINKS_INLINK_REGEX'				=> $config['primelinks_inlink_regex'],
 			'PRIMELINKS_EXLINK_REGEX'				=> $config['primelinks_exlink_regex'],
 			'PRIMELINKS_SKIP_PREFIX_REGEX'			=> $config['primelinks_skip_prefix_regex'],
+			'PRIMELINKS_INLINK_USE_TITLES'			=> $config['primelinks_inlink_use_titles'],
 			'U_ACTION'								=> $this->u_action,
 		));
 	}